backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.3
pytest-asyncio==0.24.0
